requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Werkzeug==3.0.1
pytz==2024.1
python-dateutil==2.8.2
email-validator==2.1.0